WareHouse/chatgpt_training_DefaultOrganization_20240825012145/requirements.txt
python-pptx==0.6.18
tkinter==8.6
